[tech] Separated viewer concerns into dedicated components
