+++ src/photos/components/AppRoute.jsx
@@ -5,7 +5,7 @@ import Layout from './Layout'
 import Timeline from '../containers/Timeline'
 import AlbumsView from '../containers/AlbumsView'
 import AlbumPhotos from '../containers/AlbumPhotos'
-import Viewer from 'viewer'
+import PhotosViewer from '../components/PhotosViewer'
 import { PhotosPicker } from '../ducks/albums'
 
 export const ComingSoon = () => <p style="margin-left: 2em">Coming soon!</p>
@@ -13,13 +13,13 @@ export const ComingSoon = () => <p style="margin-left: 2em">Coming soon!</p>
 const AppRoute = (
   <Route component={Layout}>
     <Route path="photos" component={Timeline}>
-      <Route path=":photoId" component={Viewer} />
+      <Route path=":photoId" component={PhotosViewer} />
     </Route>
     <Route path="albums" component={AlbumsView}>
       <Route path="new" component={PhotosPicker} />
       <Route path=":albumId" component={AlbumPhotos}>
         <Route path="edit" component={PhotosPicker} />
-        <Route path=":photoId" component={Viewer} />
+        <Route path=":photoId" component={PhotosViewer} />
       </Route>
     </Route>
     <Redirect from="/*" to="photos" />
